fixes janky camera zoom in (zoom out is still kinda weird); cleans up player input code inside of main.rs; adds a bunch of comments

+++ src/structs/ent.rs
@@ -3,7 +3,7 @@ use vector2d::Vector2D;
 
 use crate::consts::helper::CURRENT_ENT_ID;
 #[derive(Copy, Clone, PartialEq, Eq, Hash)]
-pub struct EntID(pub u32);
+pub struct EntID(pub u64);
 
 pub struct Ent {
     pub id: EntID,
